Move cypress test outside of testcase folder + revert yarn.lock
