+++ cypress/integration/2.3-add-pickup-points.spec.js
@@ -1,4 +1,4 @@
-import { testSetup, updateRetry } from '../../support/common/support'
+import { testSetup, updateRetry } from '../support/common/support'
 import {
   addPickUpPoint,
   verifyAllPickUpPoint,
